agents1/requirements.txt
crewai==0.102.0
fastapi==0.115.11
langchain_community==0.3.18
langfuse==2.59.7
pinecone==6.0.1
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
sympy==1.13.3